Don't automatically put things in navbar. Let navbar.md do its job.
Some recent pull requests added .frogrc options to disable or modify
the appearance of the automatic title and home links in the navbar. Of
course people will want to tweak these. However a simpler way to do
this is for people to use navbar.md.

As far as the title link maybe wanting a `class="brand"` attribute for
Bootstrap? People can even do that in navbar.md, because Markdown can
include HTML. So instead of

    - [My Awesome Blog](/)
    - [About](/About.html)

if they really need the "brand" classs, they can say

    - <a href="/" class="brand">My Awesome Blog</a>
    - [About](/About.html)

So there's really no need for Frog to be "helpful" and add a title
link automagically.

To be clear, this is my fault: Once I added the navbar.md feature, I
should have disabled the automatic navbar stuff. In other words, this
commit is doing what I should have, then, as well as backing out the
well-intentioned efforts to work around my neglecting to do it.