Link to documentation from Racket code blocks.
Having added support for Scribble files, Markdown is envious that
Scribble code blocks can link to documentation.

In Scribble, you get the links as a result of doing a
`(require (for-label ...))` of the module(s) you want to be used. One
the one hand it's annoying when you forget to do this. On the other
hand, when a function is provided by multiple modules --
e.g. `rectangle` or `printf` -- you control _which_ lib's
documentation is linked to.

In Markdown, you don't do this. Instead `doc-uri.rkt` has a heuristic
to pick the lib automatically. If provided by exactly one lib, of
course use that. If provided by two or more, and one of them is racket
or racket/base, use that (handles e.g. `printf`). Otherwise don't link
at all (the fate of e.g. `rectangle`).