Add .frogc option `index-newest-first?`.
Default: True. When false, posts are indexed oldest first.

Truthfully, this is a hack to let me do my presentation about Frog at
RacketCon, using RacketCon. This way, each post is a "slide", and they
will appear in the order I create them. But possibly, this might be
useful for a non-traditional style of blog (I try to convince myself).