Modify Frog to specify src and dst dirs
As-is, Frog _must_ look for a top-level directory named `_src`, and _must_
put its generated output at the top-level. This is means you have many of
your site's sources (js, css, favicon, etc.) must sit at the toplevel
with your generated output.

This change makes the source and destinations into parameters that can
be set from the config file. While the default values are the same
(`_src` and toplevel, for input and output, respectively), you can now
specify a directory for Frog's input and output, if you'd like to rename
"_src," or have the generated files go anywhere but the toplevel. See
`src-dir` and `output-dir` in `.frogrc` for more.