Handle xref-tag->path+anchor failing.
It can fail with this error:

    path-element->string: contract violation
      expected: path?
      given: 'up

which can appear for docs for a third-party library. I will loop back
and see if I can figuure out why it's failing and make a better
solution, but for now "hot fix" it to catch exception and return #f as
if the doc link were not found.