Parse `Title` metadata as markdown, sometimes. Closes #62.
When the post title will be used in certain situation -- in post index
pages, and the `@title`, `@older-title`, and `@newer-title` variables
given to `post-template.html` -- parse it as markdown, in the
expectation it will be used as the visible contents of a heading or
button element.

In all other situations -- including e.g. the `@title` variable given
to `page-template.rkt` -- keep it as plain text, in the expectation it
will be used in HTML context like <title> elements.

Update one of the example project blog posts to demonstrate this.