Initial support for rerooted blogs.
Add a uri-prefix configuration variable, which
corresponds to current-uri-prefix in the code.

A new function in paths.rkt, canonicalize-uri:
this is responsible for converting (the path
component of) a URI from internal to external form,
currently by possibly adding a prefix.

Call this function at the appropriate places (or
at least most of them).