Fix bad syntax highlighting
Suppose the text passed to `pygment-code` is following:

```
-  RPROMPT='${editor_info[overwrite]}%(?:: %F{1}⏎%f)${VIM:+" %B%F{6}V%f%b"}'
+  RPROMPT='[%D{%L:%M:%S %p}]'
+  #RPROMPT='${editor_info[overwrite]}%(?:: %F{1}⏎%f)${VIM:+" %B%F{6}V%f%b"}'
```

Obviously, we should be able to highlight it. However, `texts` will not be all string. For example, `⏎` will result in a code (number) 9166. `'` will result in `'rsquo`. Therefore, the `match` fails, and it falls back to the `[x x]` case which is definitely not what we want.

If we have something in the following form: `<pre class="brush: ...">`, it's pretty safe to assume that things inside is highlightable (if not, I feel that the original code will break too anyway!), so it's safe to apply `xexpr->markdown`.

Potentially fix #165 (I'm not exacrly sure if this is the issue that @LeifAnderson experienced or not)