Multiple authors
Closes #162 #163 #164

Allow users to specify per-post authors using new `Authors:` metadata.

Internally, authors are implemented as tags -- with an "Author: "
prefix -- so that there is a common path for features like index pages
and feeds that should support both tag and authors.

When generating the `<author>` element in Atom and RSS feeds, any
per-post authors override the blog-wide default `current-author`
setting in `.frogrc`.

Supply a new `@authors` template variable to `index-template.html` and
`post-template.html`, and update the examples to use it. (Users with
single-author blogs might want to remove this because putting "By
Me!!" on every single post might be weird. However I think it's fine
to include in the default examples, because deleting a <p> element
from a template is an easy customization to make.)

Also, warn about and ignore unknown metadata keys.