WIP blog.rkt not .frogrc
User configures using a blog.rkt file.

For existing projects, do a one-time creation of blog.rkt based on
existing .frogrc.

A future commit ought to delete some parameters that could instead be
supplied as simple function arguments in blog.rkt. This will make the
auto creation of blog.rkt a bit more complicated, which is why it's
not attempted in this commit.

Another future commit will need to update the Scribble docs.