Fix older/newer-links for BS4 as well as 3
There is now a keyword argument, for-bs, which can be either 3
(defultly) or 4 (if specified) which controls what HTML is generated.

This is kind of a horrible hack as they presumably will change it
incompatibly again in the future.  But the solution to that is to stop
them gratuitously breaking interfaces like that.  Nevertheless, I
don't know if this is the right fix.

I've changed the example post-template.html to specify BS4.  Existing
templates should continue to work for BS3, so existing blogs will not
explode.

There is no error handling for values which aren't 3 or 4: you just
get a contract violation (which is, perhaps, fine).