Don't lose listiness after unquoting
When `v` here contains a list (as opposed to a simple datum like a
string) it would, upon unquoting the entire expression, go from
something like `'(a b)` to `(a b)`.  This would then cause the error
that `a` is undefined or not a function or whatever, because the `let`
form tries to evaluate it.  Adding this quote in between quasi
quote-and-unquoting seems to fix the issue.