Prefer using setup/getinfo plus more fallbacks; fixes #253
Try to find the 'version from the package info.rkt in a variety of
ways. As the final fallback, say that the version can't be found.

I suspect that ultimate fallback might be used in the scenario for
issue #253. Why? The _package_ info.rkt data might not even exist for
a prebuilt package (maybe just one for the _collection_)?

I think a generic "version not found" message (instead of an abend)
will suffice for the issue #253 scenario, where someone is using Frog
via CI. Justification: What matters most is that it not abend
unnecessarily -- not that some version number is printed.