Increase evaluator memory limit in Frog Scribble example
This seems to be failing on stable Racket (currently 8.18) but not on
older 6.9 or on newer snapshot.

I have no idea why, I'm just trying to deal with PR #262 even though I
haven't actively maintaining Frog in quite a few years. This commit
just bumps `sandbox-memory-limit`, in a guess/hope that will work,
with no curiosity why.